Substitute and align interaction
Introduce the AlignedPlatformRule to apply substitution rules to all aligned artifacts when the substitutions are applicable. (Note: this requires that the artifact originally mentioned in the substitution rule is actually in the dependency graph). This will apply to direct and transitive dependencies.

This also takes dependency recommendations into consideration, and will provide a suggested version to replace a problematic recommendation.