Add a test use-case where there are non-locked resolution rules and locked dependencies to see the errors due to locked state out of date.
This is a rare use case, as resolution rules are typically distributed as a jar, but it can occur. In this case, core alignment takes precedence over the Nebula locking mechanism of useTarget/useVersion and the DependencyResolutionVerifier instead fails with an error that the Dependency lock state is out of date. This is introduced in the gradle-dependency-lock-plugin 11.1.+

This commit also updates the project to use gradle-dependency-lock-plugin 11.1.0-rc.1