Prevent eager task configuration resolution by avoiding dependencySubstitutions.all
`all` sets `rulesMayAddProjectDependency` to true, which means that that configuration must be resolved eagerly during task graph calculation to ensure that any project depenency substitution has occured. `eachDependency` uses `allWithDependencyResolveDetails` which doesn't set that flag, but we do pay a minor cost of adapting the ModuleVersionSelector to a ModuleComponentSelector because `DependencyResolveDetails.useTarget` won't take one directly.